+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This section contains changes that have been committed but not yet released.
 
 ### Added
 
+- Calendar Availability support
 - Add support for modifying `Folder`
 - Add support for Scheduler API